Renaming Notice to Notification
Longer, but what can we do...
