--- BTCPayServer/Views/Notice/_ViewImports.cshtml
@@ -1,1 +1,0 @@
-﻿@using BTCPayServer.Models.NoticeViewModels
