+++ BTCPayServer/Views/Notification/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model BTCPayServer.Models.NoticeViewModels.IndexViewModel
+﻿@model BTCPayServer.Models.NotificationViewModels.IndexViewModel
 @{
     ViewData["Title"] = "Notices";
 }
